add new migration

+++ database/migrations/2024_12_10_151001_create_s_e_o_metadata_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('s_e_o_metadata', function (Blueprint $table) {
             $table->id();
             $table->enum('entity_type',['product','category','blog','static_page']);
-            $table->integer('entity_id');
+            $table->unsignedBigInteger('entity_id');
             $table->string('meta_title')->nullable();
             $table->text('meta_description')->nullable();
             $table->json('meta_keywords')->nullable();
